+++ twilight-model/src/channel/message/mod.rs
@@ -1,5 +1,5 @@
 //! Textual user communication method.
-#![deny(missing_docs)]
+#![warn(missing_docs)]
 
 pub mod component;
 pub mod embed;
@@ -48,7 +48,7 @@ use crate::{
 use serde::{Deserialize, Serialize};
 
 /// Text message sent in a [`Channel`].
-#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 pub struct Message {
     /// Present with Rich Presence-related chat embeds.
     #[serde(skip_serializing_if = "Option::is_none")]