+++ twilight-http/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.15.1] - 2023-02-26
+
+### Features
+
+- get thread members request pagination ([#2119](https://github.com/twilight-rs/twilight/issues/2119))
+- add `MessageFlags::SUPPRESS_NOTIFICATIONS` ([#2129](https://github.com/twilight-rs/twilight/issues/2129))
+- support automod custom messages ([#2161](https://github.com/twilight-rs/twilight/issues/2161))
+
+### Refactor
+
+- deprecate getreactions::exec ([#2135](https://github.com/twilight-rs/twilight/issues/2135))
+- change deny lints to warn ([#2144](https://github.com/twilight-rs/twilight/issues/2144))
+
 ## [0.15.0] - 2023-02-05
 
 ### Bug Fixes